requirements.txt
pytest~=8.0.1
requests~=2.31.0
selene==2.0.0rc7
curlify~=2.2.1
allure-pytest==2.13.2
jsonschema~=4.21.1
python-dotenv==1.0.1
pydantic==2.5.2
Appium-Python-Client~=3.1.1
selenium~=4.18.1